2-5-2021

+++ lib/core/model/order_line_detail_model.dart
@@ -5,7 +5,7 @@ class OrderLineDetailModel {
   String strItemName;
   String strItemDescription;
   int intContractId;
-  double decQty;
+  int decQty;
   bool bisAlive;
   String dteCreatedDate;
   int intCreatedBy;
